connmgr: Separate mock conn and addr test code.
This moves the code related to mock addresses, connections, and
listeners to a separate file.  This helps keep it from cluttering up the
main test code and also makes it easier to reuse in other packages.
