+++ internal/connmgr/mockconn_test.go
@@ -1,0 +1,120 @@
+// Copyright (c) 2016 The btcsuite developers
+// Copyright (c) 2019-2026 The Decred developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+package connmgr
+
+import (
+	"context"
+	"errors"
+	"io"
+	"net"
+	"time"
+)
+
+// mockAddr mocks a network address.
+type mockAddr struct {
+	net, address string
+}
+
+func (m mockAddr) Network() string { return m.net }
+func (m mockAddr) String() string  { return m.address }
+
+// mockConn mocks a network connection by implementing the net.Conn interface.
+type mockConn struct {
+	io.Reader
+	io.Writer
+	io.Closer
+
+	// local network, address for the connection.
+	lnet, laddr string
+
+	// remote network, address for the connection.
+	rAddr net.Addr
+}
+
+// LocalAddr returns the local address for the connection.
+func (c mockConn) LocalAddr() net.Addr {
+	return &mockAddr{c.lnet, c.laddr}
+}
+
+// RemoteAddr returns the remote address for the connection.
+func (c mockConn) RemoteAddr() net.Addr {
+	return &mockAddr{c.rAddr.Network(), c.rAddr.String()}
+}
+
+// Close handles closing the connection.
+func (c mockConn) Close() error {
+	return nil
+}
+
+func (c mockConn) SetDeadline(t time.Time) error      { return nil }
+func (c mockConn) SetReadDeadline(t time.Time) error  { return nil }
+func (c mockConn) SetWriteDeadline(t time.Time) error { return nil }
+
+// mockDialer mocks the net.Dial interface by returning a mock connection to
+// the given address.
+func mockDialer(ctx context.Context, network, addr string) (net.Conn, error) {
+	r, w := io.Pipe()
+	c := &mockConn{rAddr: &mockAddr{network, addr}}
+	c.Reader = r
+	c.Writer = w
+	return c, ctx.Err()
+}
+
+// mockListener implements the net.Listener interface and is used to test
+// code that deals with net.Listeners without having to actually make any real
+// connections.
+type mockListener struct {
+	localAddr   string
+	provideConn chan net.Conn
+}
+
+// Accept returns a mock connection when it receives a signal via the Connect
+// function.
+//
+// This is part of the net.Listener interface.
+func (m *mockListener) Accept() (net.Conn, error) {
+	for conn := range m.provideConn {
+		return conn, nil
+	}
+	return nil, errors.New("network connection closed")
+}
+
+// Close closes the mock listener which will cause any blocked Accept
+// operations to be unblocked and return errors.
+//
+// This is part of the net.Listener interface.
+func (m *mockListener) Close() error {
+	close(m.provideConn)
+	return nil
+}
+
+// Addr returns the address the mock listener was configured with.
+//
+// This is part of the net.Listener interface.
+func (m *mockListener) Addr() net.Addr {
+	return &mockAddr{"tcp", m.localAddr}
+}
+
+// Connect fakes a connection to the mock listener from the provided remote
+// address.  It will cause the Accept function to return a mock connection
+// configured with the provided remote address and the local address for the
+// mock listener.
+func (m *mockListener) Connect(addr net.Addr) {
+	m.provideConn <- &mockConn{
+		laddr: m.localAddr,
+		lnet:  "tcp",
+		rAddr: addr,
+	}
+}
+
+// newMockListener returns a new mock listener for the provided local address
+// and port.  No ports are actually opened.
+func newMockListener(localAddr string) *mockListener {
+	return &mockListener{
+		localAddr:   localAddr,
+		provideConn: make(chan net.Conn),
+	}
+}
